Create Python DataSet Analysis
This is a project based on the FIFA World Cup 2019 and analyzes the Performance and Efficiency of Teams, Players, Countries and other related things using Python libraries like NumPy, Pandas, Matplotlib and Seaborn.
The data contains details for over 18,000 players playing in various football clubs in Europe. It contains information on age, skill rating, wages and player value, etc.